% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icerplot.R
\name{icerplot}
\alias{icerplot}
\title{Plot ICERs on a cost-effectiveness plane with incremental effectiveness and incremental cost plotted on the x and y axes, respectively.}
\usage{
icerplot(data, x, y, names = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{A dataframe containing at least two numeric columns of identical length.}

\item{x}{A variable in the dataframe containing only numeric values. Input incremental effectiveness (i.e., QALY) for ICER plot.}

\item{y}{A variable in the dataframe containing only numeric values. Input incremental cost for ICER plot.}

\item{names}{An argument to include strategy/intervention names from dataframe in the plot legend. Set to NULL as default. If NULL, labels will default to random letters in alphabetical order based on nrows in dataframe.}

\item{na.rm}{An argument specifying how to handle NA values in data before creating a plot; set to "TRUE" as default.}
}
\value{
A scatter plot with a figure legend.
}
\description{
This function generates a cost-effectiveness scatter plot intended for ICERs. It handles a dataframe with x (QALY) and y (Cost) numeric columns and an optional names column .
}
\examples{
# Creating a dataframe with ICERs
icers <- data.frame(
intervention = c("int1", "int2", "int3", "int4", "int5"),
effect = c(1, 2, 3, 3, 4),
cost = c(2, 4, 6, 2, 8))

# Applying function to create ICER plot with no "names" argument
icerplot(data = icers, x = "effect", y = "cost")

# Applying function with "names" argument specified
icerplot(data = icers, x = "effect", y = "cost", names = "intervention")
}
